Add files via upload
Kyphosis is a spine deformity and the project aims to predict the presence or absence of the disease. 
Decision Tree Classifier used here is from my own library else one can use sklearn.tree.DecisionTreeClassifier.
The algorithm is unable to handle imbalanced data very well. A random forest would perhaps be a better approach given the breadth of trees or number of estimators which remains to be tried on this dataset for me yet.
Look forward to solutions for handling imbalanced data. I've made one which sub-samples the higher number class to the lower class but it still needs to be applied to this dataset in the future. Let's hope for it and for a better result especially for a lesser bias towards the class with lesser samples.